+++ tests/openassetio-core/managerApi/HostSessionTest.cpp
@@ -6,9 +6,11 @@
 
 #include <openassetio/managerApi/HostSession.hpp>
 
+OPENASSETIO_FWD_DECLARE(LoggerInterface)
 OPENASSETIO_FWD_DECLARE(managerApi, Host)
 
 SCENARIO("HostSession constructor is private") {
-  STATIC_REQUIRE_FALSE(std::is_constructible_v<openassetio::managerApi::HostSession,
-                                               openassetio::managerApi::HostPtr>);
+  STATIC_REQUIRE_FALSE(
+      std::is_constructible_v<openassetio::managerApi::HostSession,
+                              openassetio::managerApi::HostPtr, openassetio::LoggerInterfacePtr>);
 }