+++ tests/python/openassetio/hostApi/test_manager.py
@@ -30,8 +30,8 @@ from openassetio.hostApi import Manager
 ## @todo Remove comments regarding Entity methods when splitting them from core API
 
 @pytest.fixture
-def manager(mock_manager_interface, mock_host_session):
-    return Manager(mock_manager_interface, mock_host_session)
+def manager(mock_manager_interface, a_host_session):
+    return Manager(mock_manager_interface, a_host_session)
 
 
 @pytest.fixture
@@ -78,14 +78,14 @@ def some_refs():
 class Test_Manager_init:
 
     def test_interface_returns_the_constructor_supplied_object(
-            self, mock_manager_interface, mock_host_session):
+            self, mock_manager_interface, a_host_session):
 
         # pylint: disable=protected-access
-        a_manager = Manager(mock_manager_interface, mock_host_session)
+        a_manager = Manager(mock_manager_interface, a_host_session)
         assert a_manager._interface() is mock_manager_interface
 
     def test_when_constructed_with_ManagerInterface_as_None_then_raises_TypeError(
-            self, mock_host_session):
+            self, a_host_session):
 
         # Check the message is both helpful and that the bindings
         # were loaded in the correct order such that types are
@@ -95,7 +95,7 @@ class Test_Manager_init:
             r"Manager\([^,]+managerApi.ManagerInterface,[^,]+managerApi.HostSession.+"
 
         with pytest.raises(TypeError, match=matchExpr):
-            Manager(None, mock_host_session)
+            Manager(None, a_host_session)
 
 
 class Test_Manager_identifier:
@@ -177,18 +177,18 @@ class Test_Manager_info:
 class Test_Manager_updateTerminology:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session):
+            self, manager, mock_manager_interface, a_host_session):
 
         method = mock_manager_interface.mock.updateTerminology
         a_dict = {"k", "v"}
         assert manager.updateTerminology(a_dict) is a_dict
-        method.assert_called_once_with(a_dict, mock_host_session)
+        method.assert_called_once_with(a_dict, a_host_session)
 
 
 class Test_Manager_settings:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session):
+            self, manager, mock_manager_interface, a_host_session):
 
         expected = {"some": "setting"}
         method = mock_manager_interface.mock.settings
@@ -196,120 +196,120 @@ class Test_Manager_settings:
 
         actual = manager.settings()
 
-        method.assert_called_once_with(mock_host_session)
+        method.assert_called_once_with(a_host_session)
         assert actual == expected
 
 
 class Test_Manager_initialize:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session):
+            self, manager, mock_manager_interface, a_host_session):
         a_dict = {"k": "v"}
 
         manager.initialize(a_dict)
 
-        mock_manager_interface.mock.initialize.assert_called_once_with(a_dict, mock_host_session)
+        mock_manager_interface.mock.initialize.assert_called_once_with(a_dict, a_host_session)
 
 
 class Test_Manager_prefetch:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session, some_refs, a_context):
+            self, manager, mock_manager_interface, a_host_session, some_refs, a_context):
 
         method = mock_manager_interface.mock.prefetch
         assert manager.prefetch(some_refs, a_context) == method.return_value
-        method.assert_called_once_with(some_refs, a_context, mock_host_session)
+        method.assert_called_once_with(some_refs, a_context, a_host_session)
 
 
 class Test_Manager_flushCaches:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session):
+            self, manager, mock_manager_interface, a_host_session):
 
         method = mock_manager_interface.mock.flushCaches
         assert manager.flushCaches() == method.return_value
-        method.assert_called_once_with(mock_host_session)
+        method.assert_called_once_with(a_host_session)
 
 
 class Test_Manager_isEntityReference:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session, some_refs):
+            self, manager, mock_manager_interface, a_host_session, some_refs):
 
         method = mock_manager_interface.mock.isEntityReference
         assert manager.isEntityReference(some_refs) == method.return_value
-        method.assert_called_once_with(some_refs, mock_host_session)
+        method.assert_called_once_with(some_refs, a_host_session)
 
 
 class Test_Manager_entityExists:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session, some_refs, a_context):
+            self, manager, mock_manager_interface, a_host_session, some_refs, a_context):
 
         method = mock_manager_interface.mock.entityExists
         assert manager.entityExists(some_refs, a_context) == method.return_value
-        method.assert_called_once_with(some_refs, a_context, mock_host_session)
+        method.assert_called_once_with(some_refs, a_context, a_host_session)
 
 
 class Test_Manager_defaultEntityReference:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session, a_context,
+            self, manager, mock_manager_interface, a_host_session, a_context,
             some_entity_trait_sets):
 
         method = mock_manager_interface.mock.defaultEntityReference
         assert manager.defaultEntityReference(some_entity_trait_sets, a_context) \
                 == method.return_value
-        method.assert_called_once_with(some_entity_trait_sets, a_context, mock_host_session)
+        method.assert_called_once_with(some_entity_trait_sets, a_context, a_host_session)
 
 
 class Test_Manager_entityName:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session, some_refs, a_context):
+            self, manager, mock_manager_interface, a_host_session, some_refs, a_context):
 
         method = mock_manager_interface.mock.entityName
         assert manager.entityName(some_refs, a_context) == method.return_value
-        method.assert_called_once_with(some_refs, a_context, mock_host_session)
+        method.assert_called_once_with(some_refs, a_context, a_host_session)
 
 
 class Test_Manager_entityDisplayNamer:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session, some_refs, a_context):
+            self, manager, mock_manager_interface, a_host_session, some_refs, a_context):
 
         method = mock_manager_interface.mock.entityDisplayName
         assert manager.entityDisplayName(some_refs, a_context) == method.return_value
-        method.assert_called_once_with(some_refs, a_context, mock_host_session)
+        method.assert_called_once_with(some_refs, a_context, a_host_session)
 
 
 class Test_Manager_entityVersion:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session, some_refs, a_context):
+            self, manager, mock_manager_interface, a_host_session, some_refs, a_context):
 
         method = mock_manager_interface.mock.entityVersion
         assert manager.entityVersion(some_refs, a_context) == method.return_value
-        method.assert_called_once_with(some_refs, a_context, mock_host_session)
+        method.assert_called_once_with(some_refs, a_context, a_host_session)
 
 
 class Test_Manager_entityVersions:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session, some_refs, a_context):
+            self, manager, mock_manager_interface, a_host_session, some_refs, a_context):
 
         method = mock_manager_interface.mock.entityVersions
 
         assert manager.entityVersions(some_refs, a_context) == method.return_value
         method.assert_called_once_with(
-            some_refs, a_context, mock_host_session, includeMetaVersions=False, maxNumVersions=-1)
+            some_refs, a_context, a_host_session, includeMetaVersions=False, maxNumVersions=-1)
         method.reset_mock()
 
         max_results = 5
         assert manager.entityVersions(
             some_refs, a_context, maxNumVersions=max_results) == method.return_value
         method.assert_called_once_with(
-            some_refs, a_context, mock_host_session,
+            some_refs, a_context, a_host_session,
             includeMetaVersions=False, maxNumVersions=max_results)
         method.reset_mock()
 
@@ -318,19 +318,19 @@ class Test_Manager_entityVersions:
             some_refs, a_context, maxNumVersions=max_results,
             includeMetaVersions=include_meta) == method.return_value
         method.assert_called_once_with(
-            some_refs, a_context, mock_host_session, includeMetaVersions=include_meta,
+            some_refs, a_context, a_host_session, includeMetaVersions=include_meta,
             maxNumVersions=max_results)
 
 
 class Test_Manager_finalizedEntityVersion:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session, some_refs, a_context):
+            self, manager, mock_manager_interface, a_host_session, some_refs, a_context):
 
         method = mock_manager_interface.mock.finalizedEntityVersion
         assert manager.finalizedEntityVersion(some_refs, a_context) == method.return_value
         method.assert_called_once_with(
-            some_refs, a_context, mock_host_session, overrideVersionName=None)
+            some_refs, a_context, a_host_session, overrideVersionName=None)
         method.reset_mock()
 
         a_version_name = "aVersion"
@@ -338,13 +338,13 @@ class Test_Manager_finalizedEntityVersion:
         assert manager.finalizedEntityVersion(
             some_refs, a_context, overrideVersionName=a_version_name) == method.return_value
         method.assert_called_once_with(
-            some_refs, a_context, mock_host_session, overrideVersionName=a_version_name)
+            some_refs, a_context, a_host_session, overrideVersionName=a_version_name)
 
 
 class Test_Manager_getRelatedReferences:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session, a_ref, an_empty_traitsdata,
+            self, manager, mock_manager_interface, a_host_session, a_ref, an_empty_traitsdata,
             an_entity_trait_set, a_context):
 
         # pylint: disable=too-many-locals
@@ -377,7 +377,7 @@ class Test_Manager_getRelatedReferences:
             assert manager.getRelatedReferences(
                 refs_arg, datas_arg, a_context) == method.return_value
             method.assert_called_once_with(
-                expected_refs_arg, expected_datas_arg, a_context, mock_host_session,
+                expected_refs_arg, expected_datas_arg, a_context, a_host_session,
                 resultTraitSet=None)
             method.reset_mock()
 
@@ -386,27 +386,27 @@ class Test_Manager_getRelatedReferences:
             one_ref, one_data, a_context, resultTraitSet=an_entity_trait_set) \
                     == method.return_value
         method.assert_called_once_with(
-            [one_ref], [one_data], a_context, mock_host_session,
+            [one_ref], [one_data], a_context, a_host_session,
             resultTraitSet=an_entity_trait_set)
 
 
 class Test_Manager_resolve:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session, some_refs,
+            self, manager, mock_manager_interface, a_host_session, some_refs,
             an_entity_trait_set, a_context):
 
         method = mock_manager_interface.mock.resolve
         assert manager.resolve(some_refs, an_entity_trait_set, a_context) \
                 == method.return_value
         method.assert_called_once_with(some_refs, an_entity_trait_set, a_context,
-                                       mock_host_session)
+                                       a_host_session)
 
 
 class Test_Manager_managementPolicy:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session, some_entity_trait_sets,
+            self, manager, mock_manager_interface, a_host_session, some_entity_trait_sets,
             a_context):
 
         data1 = TraitsData()
@@ -420,26 +420,26 @@ class Test_Manager_managementPolicy:
         actual = manager.managementPolicy(some_entity_trait_sets, a_context)
 
         assert actual == expected
-        method.assert_called_once_with(some_entity_trait_sets, a_context, mock_host_session)
+        method.assert_called_once_with(some_entity_trait_sets, a_context, a_host_session)
 
 
 class Test_Manager_preflight:
 
     def test_wraps_the_corresponding_method_of_the_held_interface(
-            self, manager, mock_manager_interface, mock_host_session, some_refs,
+            self, manager, mock_manager_interface, a_host_session, some_refs,
             an_entity_trait_set, a_context):
 
         method = mock_manager_interface.mock.preflight
         assert manager.preflight(some_refs, an_entity_trait_set, a_context) \
                 == method.return_value
         method.assert_called_once_with(some_refs, an_entity_trait_set, a_context,
-                                       mock_host_session)
+                                       a_host_session)
 
 
 class Test_Manager_register:
 
     def test_wraps_the_the_held_interface_register_methods(
-            self, manager, mock_manager_interface, mock_host_session, some_refs, a_traitsdata,
+            self, manager, mock_manager_interface, a_host_session, some_refs, a_traitsdata,
             a_context):
 
         datas = [a_traitsdata for _ in some_refs]
@@ -450,7 +450,7 @@ class Test_Manager_register:
 
         assert manager.register(some_refs, datas, a_context) \
                 == register_method.return_value
-        register_method.assert_called_once_with(some_refs, datas, a_context, mock_host_session)
+        register_method.assert_called_once_with(some_refs, datas, a_context, a_host_session)
 
 
     def test_when_called_with_mixed_array_lengths_then_IndexError_is_raised(
@@ -477,7 +477,7 @@ class Test_Manager_register:
 class Test_Manager_createContext:
 
     def test_context_is_created_with_expected_properties(
-            self, manager, mock_manager_interface, mock_host_session):
+            self, manager, mock_manager_interface, a_host_session):
 
         state_a = managerApi.ManagerStateBase()
         mock_manager_interface.mock.createState.return_value = state_a
@@ -488,13 +488,13 @@ class Test_Manager_createContext:
         assert context_a.retention == Context.kTransient
         assert context_a.managerState is state_a
         assert context_a.locale is None
-        mock_manager_interface.mock.createState.assert_called_once_with(mock_host_session)
+        mock_manager_interface.mock.createState.assert_called_once_with(a_host_session)
 
 
 class Test_Manager_createChildContext:
 
     def test_when_called_with_parent_then_props_copied_and_createState_called_with_parent_state(
-            self, manager, mock_manager_interface, mock_host_session):
+            self, manager, mock_manager_interface, a_host_session):
 
         state_a = managerApi.ManagerStateBase()
         mock_manager_interface.mock.createState.return_value = state_a
@@ -515,7 +515,7 @@ class Test_Manager_createChildContext:
         assert context_b.retention == context_a.retention
         assert context_b.locale == context_b.locale
         mock_manager_interface.mock.createChildState.assert_called_once_with(
-            state_a, mock_host_session)
+            state_a, a_host_session)
         mock_manager_interface.mock.createState.assert_not_called()
 
     def test_when_called_with_parent_with_no_managerState_then_createChildState_is_not_called(
@@ -535,7 +535,7 @@ class Test_Manager_createChildContext:
 class Test_Manager_persistenceTokenForContext:
 
     def test_when_called_then_the_managers_persistence_token_is_returned(
-             self, manager, mock_manager_interface, mock_host_session):
+             self, manager, mock_manager_interface, a_host_session):
 
         expected_token = "a_persistence_token"
         mock_manager_interface.mock.persistenceTokenForState.return_value = expected_token
@@ -549,7 +549,7 @@ class Test_Manager_persistenceTokenForContext:
         assert actual_token == expected_token
 
         mock_manager_interface.mock.persistenceTokenForState.assert_called_once_with(
-            initial_state, mock_host_session)
+            initial_state, a_host_session)
 
     def test_when_no_state_then_return_is_empty_and_persistenceTokenForState_is_not_called(
             self, manager, mock_manager_interface):
@@ -563,7 +563,7 @@ class Test_Manager_persistenceTokenForContext:
 class Test_Manager_contextFromPersistenceToken:
 
     def test_when_called_then_the_managers_restored_state_is_set_in_the_context(
-             self, manager, mock_manager_interface, mock_host_session):
+             self, manager, mock_manager_interface, a_host_session):
 
         expected_state = managerApi.ManagerStateBase()
         mock_manager_interface.mock.stateFromPersistenceToken.return_value = expected_state
@@ -574,7 +574,7 @@ class Test_Manager_contextFromPersistenceToken:
         assert a_context.managerState is expected_state
 
         mock_manager_interface.mock.stateFromPersistenceToken.assert_called_once_with(
-                a_token, mock_host_session)
+                a_token, a_host_session)
 
     def test_when_empty_then_no_state_and_stateFromPersistenceToken_is_not_called(
             self, manager, mock_manager_interface):